+++ src/bundle.ts
@@ -21,7 +21,6 @@ export interface BatchReadWriteRequest {
     id: string;
     vid?: string;
     resource: any;
-    ttlInSeconds?: number; // archive timestamp in Unix epoch time format in seconds
     // GET requests, only contains the URL of the resource
     fullUrl?: string;
     references?: Reference[];
@@ -30,10 +29,12 @@ export interface BatchReadWriteRequest {
 export interface BatchRequest {
     requests: BatchReadWriteRequest[];
     startTime: Date;
+    tenantId?: string;
 }
 export interface TransactionRequest {
     requests: BatchReadWriteRequest[];
     startTime: Date;
+    tenantId?: string;
 }
 
 // TODO all required?